requirements.txt
absl-py==1.1.0
antlr4-python3-runtime==4.9.3
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.5.18.1
click==8.1.3
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.7.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
hydra-core==1.2.0
importlib-metadata==4.11.4
joblib==1.1.0
Markdown==3.3.7
mkl-fft==1.3.1
mkl-service==2.4.0
oauthlib==3.2.0
omegaconf==2.2.2
packaging==21.3
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyYAML==6.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
sklearn==0.0
smmap==5.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.5.4
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
wandb==0.12.18
Werkzeug==2.1.2
zipp==3.8.0
